refactor: split package.go into one file per platform
package.go (config/gates/setup guide/shared helpers) + package_ios_ipa.go,
package_android_aab.go, package_macos_app.go, package_windows_msix.go.
Named with non-GOOS suffixes (_ios_ipa etc.) so Go doesn't treat them as
platform-constrained build files.
